backend/requirements.txt
Django
djangorestframework
django-rest-swagger
drf-yasg
django-cors-headers
yfinance --upgrade --no-cache-dir
pandas
pandas-plotly







